Administration (#39)
* Beginning of administration

* Server side administration!

* Beginning of administration

* Server side administration!

* Fix

* Fix logout & notify client on token perm change

* Notify when promoted too

* Add admin settings button if user is admin at login

Still needs to be done when server updates perms

* Update admin perms on change

* forgot to commit file

* Accessible settings!

* Fix

* Styling

* Search users

* Fmt & clippy

* Search for users! needs styling & a11y

* List all users

* Use treeview

* Use treeview + style

* Ban users

* Unban users

* Unlock users

* Add reports table code

* small fixes

* Commit

* commit

* commit

* Style

* System uri

* Better style and such

* add queries for reports

* Report users

* Search and take action on reports

* Remove debug print

* Set compromised

* fix

* Fix mime type for handling vertex

Still needs to have windows support

* Set admin perms when boxes ticked

* High contrast css

+++ server/src/database/token.rs
@@ -90,7 +90,7 @@ impl Database {
             &token.device.0,
             &token.device_name,
             &token.token_hash,
-            &(token.hash_scheme_version as u8 as i16),
+            &(token.hash_scheme_version as i16),
             &token.user.0,
             &token.last_used,
             &token.expiration_date,
